+++ src/index.d.ts
@@ -42,7 +42,46 @@ declare namespace maplibregl {
      * Tiles may still be cached by the browser in some cases.
      */
     export function clearStorage(callback?: (err?: Error) => void): void;
-    
+    /**
+     * Sets a custom load tile function that will be called when using a source that starts with a custom url schema.
+     * The example below will be triggered for custom:// urls defined in the sources list in the style definitions.
+     * The function passed will receive the request parameters and should call the callback with the resulting request,
+     * for example a pbf vector tile, non-compressed, represented as ArrayBuffer.
+     * @param {string} customProtocol - the protocol to hook, for example 'custom'
+     * @param {Function} loadFn - the function to use when trying to fetch a tile specified by the customProtocol
+     * @example
+     * // this will fetch a file using the fetch API (this is obviously a non iteresting example...)
+     * maplibre.addProtocol('custom', (params, callback) => {
+            fetch(`https://${params.url.split("://")[1]}`)
+                .then(t => {
+                    if (t.status == 200) {
+                        t.arrayBuffer().then(arr => {
+                            callback(null, arr, null, null);
+                        });
+                    } else {
+                        callback(new Error(`Tile fetch error: ${t.statusText}`));
+                    }
+                })
+                .catch(e => {
+                    callback(new Error(e));
+                });
+            return { cancel: () => { } };
+        });
+     * // the following is an example of a way to return an error when trying to load a tile
+     * maplibre.addProtocol('custom2', (params, callback) => {
+     *      callback(new Error('someErrorMessage'));
+     *      return { cancel: () => { } };
+     * });
+     */
+    export function addProtocol(customProtocol: string, loadFn: (requestParameters: RequestParameters, callback: ResponseCallback<any>) => Cancelable);
+    /**
+     * Removes a previusly added protocol
+     * @param {string} customProtocol - the custom protocol to remove registration for
+     * @example
+     * maplibregl.removeProtocol('custom');
+     */
+    export function removeProtocol(customProtocol: string);
+
     export function setRTLTextPlugin(pluginURL: string, callback: (error: Error) => void, deferred?: boolean): void;
     export function getRTLTextPluginStatus(): PluginStatus;